requirement.txt
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Flask
Flask-Cors==1.10.3
fonttools==4.39.3
idna==3.4
itsdangerous 
Jinja2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe
matplotlib==3.7.1
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
pysolr==3.9.0
pyspellchecker==0.7.1
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
scipy==1.10.1
six==1.16.0
textblob==0.17.1
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug
